feat: Implement WireframeRenderer in C++ and expose to Python
- Added WireframeRenderer class for efficient wireframe rendering using pre-built unit meshes (circle, arc, box, line).
- Introduced helper functions for matrix operations (identity, translation, scaling, rotation).
- Replaced Python implementation of wireframe rendering with C++ for performance improvements.
- Updated SDL embedded backend to support framebuffer blitting for debugging.
- Modified ColorPass to utilize C++ implementation, improving rendering efficiency and reducing Python overhead.
- Cleaned up and optimized collider gizmo rendering logic.

+++ termin/visualization/render/framegraph/passes/collider_gizmo.py
@@ -100,8 +100,8 @@ class ColliderGizmoPass(RenderFramePass):
         graphics.bind_framebuffer(fb)
         graphics.set_viewport(0, 0, pw, ph)
 
-        view = camera.get_view_matrix()
-        proj = camera.get_projection_matrix()
+        view = camera.get_view_matrix().astype(np.float32)
+        proj = camera.get_projection_matrix().astype(np.float32)
 
         self._renderer.begin(graphics, view, proj, depth_test=False)
         self._draw_colliders(scene)
